requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Pillow==7.2.0
PyExifTool==0.1.1
requests==2.24.0
urllib3==1.25.10
lambda-cache==0.8.1
